+++ superduper/backends/ibis/data_backend.py
@@ -167,10 +167,10 @@ class IbisDataBackend(BaseDataBackend):
         fields = {
             INPUT_KEY: dtype('string'),
             'id': dtype('string'),
-            f'_outputs.{predict_id}': output_type,
+            f'_outputs__{predict_id}': output_type,
         }
         return Table(
-            identifier=f'_outputs.{predict_id}',
+            identifier=f'_outputs__{predict_id}',
             schema=Schema(identifier=f'_schema/{predict_id}', fields=fields),
         )
 
@@ -180,7 +180,7 @@ class IbisDataBackend(BaseDataBackend):
         :param predict_id: The identifier of the prediction.
         """
         try:
-            self.conn.table(f'_outputs.{predict_id}')
+            self.conn.table(f'_outputs__{predict_id}')
             return True
         except (NoSuchTableError, ibis.IbisError):
             return False