+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add helper dict method in Event.
 - Use declare_component from base class.
 - Use different colors to distinguish logs
+- Change all the `_outputs.` to `_outputs__`
 
 
 #### New Features & Functionality